[Refactor] Deduplicate UnionFind utilities (#27611)
So many UnionFild utilities in code base, so unify it into one
