+++ fe/fe-core/src/main/java/com/starrocks/common/util/UnionFind.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.starrocks.sql.optimizer;
+package com.starrocks.common.util;
 
 import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
